Refactor test according to ashs comments.
